omc: Initial implementation of entity sets omc: Implementation of function-style cast operators in expression attributes omc: re-work agentvar design to resolve overload issues omc: miscellaneous fixes for IDMM IDMM: require explicit cast in model model code when assigning type int, etc. to model types (e.g. ranges)

+++ include/omc/Multilink.h
@@ -146,7 +146,7 @@ public:
         link<B>	next_lnk = nullptr;
 
         *next_pos = -1;
-        for ( index = start; index < storage.size(); ++index ) {
+        for ( index = start; index < (int)storage.size(); ++index ) {
             lnk = storage[index];
             if ( lnk != nullptr ) {
                 next_lnk = lnk;
